--- src/sort/mod.rs
@@ -1,5 +1,0 @@
-use egglog::EGraph;
-use ruler::Symbol;
-
-pub mod option;
-pub mod rule;
