--- tests/egglog/math-predicates.egg
@@ -1,9 +1,0 @@
-(datatype PredFn
-   (Eq)
-   (Neq)
-   (Lt)
-   (Gt))
-
-(datatype Pred
-   (PredOp2 PredFn Math Math))
-
